requirements.txt
flask
flask-cors
pillow
ultralytics